Limit it to 16bit.

+++ router.h
@@ -37,9 +37,9 @@ struct router {
     rtime_t *best_time;
 
     /* The location of the state in the smaller malloc */
-    uint32_t *states_pointer;
+    uint16_t *states_pointer;
 
-    uint32_t n_states;
+    uint16_t n_states;
 
    /* The index of the journey_pattern used to travel from back_stop to here, or WALK  */
     uint32_t *states_back_journey_pattern;
